Fixed redirect error

+++ interfaces/src/main/java/ar/edu/itba/paw/persistence/ArtistDao.java
@@ -11,7 +11,7 @@ public interface ArtistDao {
     Optional<Artist> findById(long id);
     List<Artist> findAll();
     List<Artist> findBySongId(long id);
-    int save(Artist artist);
+    long save(Artist artist);
     int update(Artist artist);
     int deleteById(long id);
 }
